Folder structure changes
Second changeset to order files as required
