--- src/chrome/manifest.json
@@ -1,20 +1,0 @@
-{
-  "manifest_version": 2,
-
-  "name": "YouTube standard/TV switcher",
-  "description": "This extension allows to switch between desktop/TV versions of the YouTube platform",
-  "version": "0.0.2",
-  "version_name": "0.0.2-alpha",
-  "browser_action": {
-    "default_icon": "icon.png"
-  },
-  "icons": { "16": "icon16.png",
-           "48": "icon48.png",
-          "128": "icon128.png" },
-  "background": {
-    "scripts": ["switcher.js"]
-  },
-  "permissions": [
-    "activeTab"
-  ]
-}
