--- src/chrome/switcher.html
@@ -1,9 +1,0 @@
-<html>
-  <head>
-    <title>YouTube switcher background page</title>
-    <script type="text/javascript" src="https://ajax.googleapis.com/ajax/libs/jquery/3.0.0/jquery.min.js"></script>
-  </head>
-  <body>
-    <script type="text/javascript" src="switcher.js"></script>
-  </body>
-</html>
