+++ src/switcher.html
@@ -1,0 +1,8 @@
+<html>
+  <head>
+    <title>YouTube switcher background page</title>
+  </head>
+  <body>
+    <script type="text/javascript" src="switcher.js"></script>
+  </body>
+</html>
